i never want to touch bra ever again
this has to be the most annoying opcode defintion to deal with
both in terms of how it operates as well as the random bs it throws out randomly 

granted, this could very well be at a fault of my own what with the fact that my own implementations arent exactly the best and quite frankly, have a lot of fallthroughs

BUT, I have extensively looked over at docs and made sure that for each size extension, it handles the offset accordingly 

WORD is always so bloody weird as it is always has to account for an extra displacement. Suppose that for 8 bit variants, the displacement is part of the opcode itself due to the limited size but 16 and 32 need to account for the additional displacement in it's extension so a -2 is required

fingers crossed that it all works and that i dont have to touch it again - more stuff to document i suppose?